+++ etils/array_types/dtypes.py
@@ -17,22 +17,12 @@
 from __future__ import annotations
 
 import abc
-import typing
-from typing import Any
+import dataclasses
+from typing import Any, ClassVar, Optional
 
-from etils import epy
+from etils.enp import numpy_utils
 import numpy as np
 
-if typing.TYPE_CHECKING:
-  from etils import enp
-
-
-# TODO(epot): Replace dtype by generic `AnyFloat`, `AnyInt` dtype.
-_STD_TYPE_TO_NP_DTYPE = {
-    int: np.int32,
-    float: np.float32,
-    bool: np.bool_,
-}
 _NP_KIND_TO_STR = {
     'u': 'ui',
     'i': 'i',
@@ -47,6 +37,17 @@ _NP_KIND_TO_STR = {
 _BITS_KINDS = {'u', 'i', 'f', 'c', 'V'}
 
 
+def _make_array_cls_name(np_dtype: np.dtype) -> str:
+  """Makes the array class name for the dtype."""
+  kind_str = _NP_KIND_TO_STR[np_dtype.kind]
+  if np_dtype.kind in _BITS_KINDS:
+    # Display with the size (ui8, f32,...)
+    return f'{kind_str}{np_dtype.itemsize * 8}'
+  else:
+    return kind_str  # Raw types (str, bool_,...)
+
+
+@dataclasses.dataclass
 class DType(abc.ABC):
   """DType wrapper.
 
@@ -54,31 +55,46 @@ class DType(abc.ABC):
 
   This is EXPERIMENTAL, so the API might change.
 
+  Attributes:
+    name: Representation name (e.g. np.uint8, AnyFloat...)
+    array_cls_name: Name of the array class associated with the dtype (`f32`,
+      `ui8`,...).
   """
+  name: str
+  array_cls_name: str
 
   @classmethod
   def from_value(cls, value: Any) -> 'DType':
     """Convert the value to dtype."""
     if value is None:
       return AnyDType()
-    if isinstance(value, DType):
+    elif isinstance(value, DType):
       return value
-    elif isinstance(value, np.dtype) or epy.issubclass(value, np.generic):
-      return NpDType(value)
-    if value in _STD_TYPE_TO_NP_DTYPE:
-      return NpDType(_STD_TYPE_TO_NP_DTYPE[value])
+    elif numpy_utils.lazy.is_dtype(value):
+      return NpDType(numpy_utils.lazy.as_dtype(value))
+    elif value in _STD_TYPE_TO_DTYPE:
+      return _STD_TYPE_TO_DTYPE[value]
     else:
       raise TypeError(f'Unsuported dtype: {value!r}')
 
-  @property
-  @abc.abstractmethod
-  def array_cls_name(self) -> str:
-    """Name of the array class associated with the dtype (`f32`, `ui8`,...)."""
-    raise NotImplementedError
+  def asarray(self, array_like, *, xnp: numpy_utils.NpModule):
+    from_dtype = _get_array_dtype(array_like)
+    to_dtype = self._get_target_dtype(from_dtype)
+    return xnp.asarray(array_like, dtype=to_dtype)
 
   @abc.abstractmethod
-  def asarray(self, array_like, *, xnp: enp.NpModule):
-    raise NotImplementedError
+  def _get_target_dtype(
+      self,
+      from_dtype: Optional[np.dtype],
+  ) -> Optional[np.dtype]:
+    """Validate and normalize the numpy dtype.
+
+    Args:
+      from_dtype: DType of the array to cast
+
+    Returns:
+      to_dtype: DType of the array after casting
+    """
 
   @abc.abstractmethod
   def __eq__(self, other) -> bool:
@@ -89,7 +105,7 @@ class DType(abc.ABC):
     raise NotImplementedError
 
   def __repr__(self) -> str:
-    return f'DType({self.array_cls_name!r})'
+    return f'DType({self.name})'
 
 
 class NpDType(DType):
@@ -97,22 +113,19 @@ class NpDType(DType):
 
   def __init__(self, dtype):
     self.np_dtype = np.dtype(dtype)  # Normalize the dtype
+    super().__init__(
+        name=str(self.np_dtype),
+        array_cls_name=_make_array_cls_name(self.np_dtype),
+    )
 
-  @property
-  def array_cls_name(self) -> str:
-    np_dtype = self.np_dtype
-    kind_str = _NP_KIND_TO_STR[np_dtype.kind]
-    if np_dtype.kind in _BITS_KINDS:
-      # Display with the size (ui8, f32,...)
-      return f'{kind_str}{np_dtype.itemsize * 8}'
-    else:
-      return kind_str  # Raw types (str, bool_,...)
-
-  def asarray(self, array_like, *, xnp: enp.NpModule):
-    # from etils import enp  # pylint: disable=g-import-not-at-top
-    # TODO(epot): Add a strict=True mode to prevent `int` -> `float` conversion
-    # Normalize list,... as array
-    return xnp.asarray(array_like, dtype=self.np_dtype)
+  def _get_target_dtype(
+      self,
+      from_dtype: Optional[np.dtype],
+  ) -> Optional[np.dtype]:
+    # Here, we could validate invalid casting, like:
+    # * float -> int
+    # * int -> bool
+    return self.np_dtype
 
   def __eq__(self, other) -> bool:
     _assert_isdtype(other)
@@ -122,29 +135,101 @@ class NpDType(DType):
     return hash(self.np_dtype)
 
 
-class AnyDType(DType):
-  """DType which can represent any dtype."""
+class _SingletonDType(DType):
+  """DType without arguments."""
 
-  @property
-  def array_cls_name(self) -> str:
-    return 'Array'
+  ARRAY_CLS_NAME: ClassVar[str]
 
-  def asarray(self, array_like, *, xnp: enp.NpModule):
-    return xnp.asarray(array_like)
+  def __init__(self):
+    super().__init__(
+        name=self.__class__.__name__,
+        array_cls_name=self.ARRAY_CLS_NAME,
+    )
 
   def __eq__(self, other) -> bool:
     _assert_isdtype(other)
-    return isinstance(other, AnyDType)
+    return isinstance(other, type(self))
 
   def __hash__(self) -> int:
     return hash(type(self).__qualname__)
 
 
-class DTypeUnion(DType):
-  """Generic dtype (float, int,...)."""
+class AnyDType(_SingletonDType):
+  """DType which can represent any dtype."""
 
-  def asarray(self, array_like, *, xnp: enp.NpModule):
-    raise NotImplementedError
+  ARRAY_CLS_NAME = 'Array'
+
+  def _get_target_dtype(
+      self,
+      from_dtype: Optional[np.dtype],
+  ) -> Optional[np.dtype]:
+    return None  # Keep original dtype
+
+
+class AnyFloat(_SingletonDType):
+  """Generic float dtype (float32, float64, bfloat16,...)."""
+
+  ARRAY_CLS_NAME = 'FloatArray'
+
+  def _get_target_dtype(
+      self,
+      from_dtype: Optional[np.dtype],
+  ) -> Optional[np.dtype]:
+    if from_dtype is None:
+      return np.float32
+    elif _is_float(from_dtype):
+      return from_dtype
+    else:  # int, bool,...
+      # Could validate dtype to restrict too implicit casting
+      return np.float32
+
+
+class AnyInt(_SingletonDType):
+  """Generic int dtype (int32, int64, uint8,...)."""
+
+  ARRAY_CLS_NAME = 'IntArray'
+
+  def _get_target_dtype(
+      self,
+      from_dtype: Optional[np.dtype],
+  ) -> Optional[np.dtype]:
+    if from_dtype is None:
+      return np.int32
+    elif _is_integer(from_dtype):
+      return from_dtype
+    else:  # float, bool,...
+      # Could validate dtype to restrict too implicit casting
+      return np.int32
+
+
+_STD_TYPE_TO_DTYPE = {
+    int: AnyInt(),
+    float: AnyFloat(),
+    bool: NpDType(np.bool_),
+}
+
+
+def _get_array_dtype(array_like) -> Optional[np.dtype]:
+  """Returns the numpy dtype from the array."""
+  if numpy_utils.lazy.is_array(array_like):  # Already an ndarray
+    return numpy_utils.lazy.as_dtype(array_like.dtype)
+  else:
+    # TODO(epot): Could have a smarter way of infering the dtype for
+    # scalar, int, float,... but difficult to infer list without performance
+    # cost (one way would be to call `asarray(array_like, dtype=None)`, then
+    # cast again)
+    return None
+
+
+def _is_float(dtype: np.dtype) -> bool:
+  """Validate the dtype is float."""
+  # `V` to support bfloat16
+  return np.issubdtype(dtype, np.floating) or dtype.kind == 'V'
+
+
+def _is_integer(dtype: np.dtype) -> bool:
+  """Validate the dtype is integer."""
+  return np.issubdtype(dtype, np.integer)
 
 
 def _assert_isdtype(dtype: Any) -> None: